Fixed our Tree so we now have a Program with a list of FunctionBlocks. And each FunctionBlock has a ParamDefList, name, and StatementList.  Added DeclStatements for declarations of variables (int x;). Added ReturnStatement for returns in functions (return x;). Added FuncCall for function calls and their return with a name and a list of Expr in a ExprList.  Doesn't do much since the AssemblyVisitor will do all of the heavy lifting.

+++ Expr.h
@@ -14,6 +14,7 @@ public:
 
     virtual void Execute()
     {
+        // could remove this probably
         Evaluate();
         std::cout << "Answer: " << *this << std::endl;
     }
@@ -22,7 +23,7 @@ public:
 
     virtual Expr* Clone() = 0;
 
-    virtual double Get() const { return _value; }
+    virtual int Get() const { return _value; }
 
     /**
      * Print an arbitrary expression to a stream
